Separate mapper tab from the mapper
This PR separates the mapper tab from the mapper implementation.
This allows compiling the mapper implementation without DOM APIs.
As the next step we can turn bidiMapper/bidiTab/bidiServer into composite projects
to reduce duplication in the build output.
