+++ .github/workflows/unit.js.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Build
         run: npm run build
       - name: Verify mapper size
-        run: du -sh lib/iife/bidiMapper/mapper.js
+        run: du -sh lib/iife/mapperTab.js
       - name: Run unit tests
         run: npm run unit
